Added in proof of concept for leveraging TitanDB multi-value properties instead of delimeters for the types.
This improvement is to fix query times; this would allow us to search for a type in O(1) time due to indexing, instead of having to do a text-scan over the index, as is currently implemented.

The major section of change is the GraphTypeManager, which adds the TitanDB Multi-value.  Also, QueryTypeCriterion has been modified to search by equals as opposed to contains.