Updated TechnologyIdentified to a generic approach and added a unit test.
Previously, getApplications() was doing the same thing as getRootProjectModels(). Also, getRootProjectModels() was returning
only the shared-libs application in the case of a jar found in multiple applications.

This update adds a unit test to confirm this behavior and also makes the code attempt to associate the technology
statististics at the lowest possible module level. This help when we (eventually) enable filtering support.